V1.0 Funciones básicas (El editor da problemas almecenando imagen)

+++ app/Http/Controllers/PostController.php
@@ -85,7 +85,8 @@ class PostController extends Controller
 
         
         if($request->hasFile('Image')){
-            /*if(Storage::disk('outLaravel')->exists($name)){
+
+            if(Storage::disk('outLaravel')->exists($name)){
                 Storage::disk('outLaravel')->delete($name);
                 $request->Image->storeAs('',$name,'outLaravel');
                 $this->Image = $name;
@@ -97,8 +98,11 @@ class PostController extends Controller
                 }else{
                     $this->Image = "Default.jpg";
                 }
-            }*/
+            }
 
+
+            /*
+                HEROKU CODIGO INTENTO
             $request->Image->storeAs('postImages',$name);
             if(Storage::disk('public')->exists($name)){
                 Storage::disk('public')->delete($name);
@@ -110,7 +114,7 @@ class PostController extends Controller
                 }else{
                     $this->Image = "Default.jpg";
                 }
-            }
+            }*/
 
         }else{
             return redirect('home');
@@ -190,7 +194,7 @@ class PostController extends Controller
         $name = "";
 
         if($request->hasFile('Image')){
-            /*$name = $request->Image->getClientOriginalName();
+            $name = $request->Image->getClientOriginalName();
             if(Storage::disk('outLaravel')->exists($name)){
                 Storage::disk('outLaravel')->delete($name);
                 $request->Image->storeAs('',$name,'outLaravel');
@@ -201,7 +205,10 @@ class PostController extends Controller
                 }else{
                     $this->Image = $lastImage->Imagen;
                 }
-            }*/
+            }
+
+            /*
+                    HEROKU CODIGO
             if(Storage::disk('public')->exists($name)){
                 Storage::disk('public')->delete($name);
                 $request->Image->storeAs('postImages',$name,'puexistsblic');
@@ -212,7 +219,7 @@ class PostController extends Controller
                 }else{
                     $this->Image = "Default.jpg";
                 }
-            }
+            }*/
         }else{
             $this->Image = $lastImage->Imagen;  
         }
@@ -244,10 +251,10 @@ class PostController extends Controller
 
         $post = post::find($id);
         
-        /*if(Storage::disk('outLaravel')->exists($post->Imagen)){
+        if(Storage::disk('outLaravel')->exists($post->Imagen)){
             Storage::disk('outLaravel')->delete($post->Imagen);
             $post->delete();
-        }*/
+        }
         $post->delete();
 
         return redirect('home');
